requirements.txt
tensorflow==2.21.0
numpy==2.4.3
scikit-learn==1.8.0
matplotlib==3.10.8
seaborn==0.13.2
pillow==12.1.1